added strip_rows to utils

+++ neural/alignment.py
@@ -300,16 +300,17 @@ def align_epi_anat(anatomy,epi_dsets,skull_strip_anat=True):
 def skullstrip_template(dset,template,dilate=0):
     '''Takes the raw anatomy ``dset``, aligns it to a template brain, and applies a templated skullstrip. Should produce fairly reliable skullstrips as long
     as there is a decent amount of normal brain and the overall shape of the brain is normal-ish'''
-    with nl.notify('Running template-based skull-strip on %s' % dset):
-        dset = os.path.abspath(dset)
-        template = os.path.abspath(template)
-        tmp_dir = tempfile.mkdtemp()
-        cwd = os.getcwd()
-        with nl.run_in(tmp_dir):
-            nl.affine_align(template,dset,skull_strip=None,cost='mi',opts=['-nmatch','100%'])
-            nl.run(['3dQwarp','-minpatch','20','-penfac','10','-noweight','-source',nl.suffix(template,'_aff'),'-base',dset,'-prefix',nl.suffix(template,'_qwarp')],products=nl.suffix(template,'_qwarp'))
-            info = nl.dset_info(nl.suffix(template,'_qwarp'))
-            max_value = info.subbricks[0]['max']    
-            nl.calc([dset,nl.suffix(template,'_qwarp')],'a*step(b-%f*0.05)'%max_value,nl.suffix(dset,'_sstemplate'))
-            shutil.move(nl.suffix(dset,'_sstemplate'),cwd)
-        shutil.rmtree(tmp_dir)
+    if not os.path.exists(nl.suffix(dset,'_sstemplate')):
+        with nl.notify('Running template-based skull-strip on %s' % dset):
+            dset = os.path.abspath(dset)
+            template = os.path.abspath(template)
+            tmp_dir = tempfile.mkdtemp()
+            cwd = os.getcwd()
+            with nl.run_in(tmp_dir):
+                nl.affine_align(template,dset,skull_strip=None,cost='mi',opts=['-nmatch','100%'])
+                nl.run(['3dQwarp','-minpatch','20','-penfac','10','-noweight','-source',nl.suffix(template,'_aff'),'-base',dset,'-prefix',nl.suffix(template,'_qwarp')],products=nl.suffix(template,'_qwarp'))
+                info = nl.dset_info(nl.suffix(template,'_qwarp'))
+                max_value = info.subbricks[0]['max']    
+                nl.calc([dset,nl.suffix(template,'_qwarp')],'a*step(b-%f*0.05)'%max_value,nl.suffix(dset,'_sstemplate'))
+                shutil.move(nl.suffix(dset,'_sstemplate'),cwd)
+            shutil.rmtree(tmp_dir)
